+++ src/Databento.CSharpApiClient/DatabentoClient.cs
@@ -48,7 +48,7 @@ namespace Databento.CSharpApiClient
             }
             else
             {
-                HttpClient httpClient = new HttpClient();
+                HttpClient httpClient = new HttpClient(HttpCompression.CreateDecompressingHandler(), disposeHandler: true);
                 httpClient.BaseAddress = this.options.BaseUri;
                 httpClient.Timeout = this.options.Timeout;
                 httpClient.DefaultRequestHeaders.UserAgent.ParseAdd(this.options.UserAgent ?? "DatabentoClient/1.0");